sagalog-requirements.txt
aiopulsar-py==0.1.0a0
fastapi==0.89.1
uvicorn[standard]==0.20.0
pulsar-client==2.10.2
pulsar-client[avro]==2.10.2
requests==2.28.2
sse-starlette==1.2.1